✨ map-api 관련 util 함수 기능 #11
